Enable Agent-VPP config msg communication (#18184)
1. Support starting VPP in mock mode without model.
2. Collect pdsagent backtrace for debugging agent hang issues.

+++ nic/apollo/test/scale/apulu/run_scale_test_mock.sh
@@ -18,15 +18,24 @@ if [[ "$1" ==  --cfg ]]; then
 fi
 
 function finish {
-   echo "===== Collecting logs ====="
-   ${NICDIR}/apollo/test/tools/savelogs.sh
-   rm -f ${NICDIR}/conf/pipeline.json
+    echo "===== Collecting logs ====="
+    ${NICDIR}/apollo/test/tools/savelogs.sh
+    rm -f ${NICDIR}/conf/pipeline.json
+    pkill -9 vpp
 }
 trap finish EXIT
 
 export PATH=${PATH}:${BUILD_DIR}/bin
 rm -f $NICDIR/conf/pipeline.json
 ln -s $NICDIR/conf/apulu/pipeline.json $NICDIR/conf/pipeline.json
+
+echo "Starting VPP"
+sudo $NICDIR/vpp/tools/start-vpp-mock.sh --pipeline apulu
+if [[ $? != 0 ]]; then
+    echo "Failed to bring up VPP"
+    exit -1
+fi
+
 $GDB apulu_scale_test -c hal.json -i ${NICDIR}/apollo/test/scale/$cfgfile --gtest_output="xml:${GEN_TEST_RESULTS_DIR}/apulu_scale_test.xml"
 rm -f $NICDIR/conf/pipeline.json
 if [ $? -eq 0 ]
